requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.6
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
importlib-resources==6.0.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
protobuf==4.23.4
pycryptodome==3.18.0
python-dotenv==1.0.0
pywin32==306
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.8.8
toolz==0.12.0
urllib3==2.0.3
web3==6.5.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
